labwork4/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
marshmallow==3.20.1
python-dotenv==1.0.0